+++ src/main/kotlin/com/github/kloostermanw/ideacelery/listeners/CeleryProjectManagerListener.kt
@@ -1,13 +1,13 @@
 package com.github.kloostermanw.ideacelery.listeners
 
-import com.github.kloostermanw.ideacelery.services.MyProjectService
+import com.github.kloostermanw.ideacelery.services.CeleryProjectService
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManagerListener
 
-internal class MyProjectManagerListener : ProjectManagerListener {
-
+internal class CeleryProjectManagerListener : ProjectManagerListener {
+    @Suppress("OVERRIDE_DEPRECATION")
     override fun projectOpened(project: Project) {
-        project.service<MyProjectService>()
+        project.service<CeleryProjectService>()
     }
 }