+++ src/main/kotlin/com/github/kloostermanw/ideacelery/services/CeleryApplicationService.kt
@@ -1,0 +1,9 @@
+package com.github.kloostermanw.ideacelery.services
+
+import com.github.kloostermanw.ideacelery.CeleryBundle
+
+class CeleryApplicationService {
+    init {
+        println(CeleryBundle.message("applicationService"))
+    }
+}