+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# idea-celery Changelog
+# intelli-celery-web-app Changelog
 
-## [Unreleased]
+## Unreleased
+
+## 0.0.1
+
 ### Added
+
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- PHP Language::get() completion contributor